move core node-file-trace logic into turbopack
