+++ turbopack/crates/turbopack/src/nft_json.rs
